create and add the LogoSection in Hero
